requirements.txt
pandas
scikit-learn
seaborn
matplotlib
numpy
joblib
pyserial
fastapi
uvicorn
sqlalchemy
passlib
